# openapi.json
{
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Uganda-first professional + dating network. Envelope: `{code:1|0, message, data}`. Auth: Bearer JWT.",
    "title": "LinkUp API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/admin/auth/login": {
      "post": {
        "operationId": "post_admin.admin_login",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "summary": "POST /api/admin/auth/login",
        "tags": [
          "api"
        ]
      }
    },
    "/api/admin/dashboard": {
      "get": {
        "operationId": "get_admin.dashboard",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /api/admin/dashboard",
        "tags": [
          "api"
        ]
      }
    },
...
    "/api/admin/users": {
      "get": {
        "operationId": "get_admin.users_index",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /api/admin/users",
        "tags": [
          "api"
        ]
      }
    },
    "/api/admin/users/{user_id}": {
      "get": {
        "operationId": "get_admin.users_show",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /api/admin/users/{user_id}",
        "tags": [
          "api"
        ]
      }
    },
    "/api/admin/users/{user_id}/delete": {
      "delete": {
        "operationId": "delete_admin.users_delete",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "DELETE /api/admin/users/{user_id}/delete",
        "tags": [
          "api"
        ]
      },
      "post": {
        "operationId": "post_admin.users_delete",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /api/admin/users/{user_id}/delete",
        "tags": [
          "api"
        ]
      }
    },
    "/api/admin/users/{user_id}/reset-password": {
      "post": {
        "operationId": "post_admin.users_reset_password",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /api/admin/users/{user_id}/reset-password",
        "tags": [
          "api"
        ]
      }
    },
    "/api/admin/users/{user_id}/toggle-status": {
      "post": {
        "operationId": "post_admin.toggle_status",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /api/admin/users/{user_id}/toggle-status",
        "tags": [
          "api"
        ]
      }
    },
    "/api/admin/users/{user_id}/update": {
      "post": {
        "operationId": "post_admin.users_update",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /api/admin/users/{user_id}/update",
        "tags": [
          "api"
        ]
      },
      "put": {
        "operationId": "put_admin.users_update",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "PUT /api/admin/users/{user_id}/update",
        "tags": [
          "api"
        ]
      }
    },
    "/api/admin/users/{user_id}/wallet": {
      "get": {
        "operationId": "get_admin.users_wallet",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /api/admin/users/{user_id}/wallet",
        "tags": [
          "api"
        ]
      }
    },
...
    "/api/wallet/transactions": {
      "get": {
        "operationId": "get_wallet.transactions",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /api/wallet/transactions",
        "tags": [
          "api"
        ]
      }
    },
...
    "/v1/admin/login": {
      "post": {
        "operationId": "post_v1_admin.admin_login",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /v1/admin/login",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/wallet/transactions": {
      "get": {
        "operationId": "get_v1_wallet.transactions",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /v1/wallet/transactions",
        "tags": [
          "wallet"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:5001"
    }
  ],
...
}